// src/app/shared/models/product.model.ts
export interface Product {
    id: number;
    name: string;
    description: string;
    price: number;
    discountPercent: number;
    discountPrice: number;
    stock: number;
    imageUrl: string;
    brand: string;
    ratings: number;
    reviewCount:any;
    categoryName: string;
    createdAt: string
}